requirements.txt
numpy
scipy
scikit-learn
smplx==0.1.28
tqdm
yacs
absl-py
ml-collections
tensorboardX
matplotlib
chumpy
opencv-python
torchgeometry
plyfile
pyglet
pymeshlab
pyrender
trimesh
pytorch-lightning
pytorch3d